minibrowser.c: add GOption based command line parsing
This patch adds a command line option "-r" that allows us to restrict
navigation to URLs matching the pattern passed as argument (the
feature added by Miguel).

Moved main() to the bottom of the file, is the most common location. Also
tweaked a bit the Makefile, now checks for changes in the source code.